// modules/medicalOptional/types/medical-optional-errors.ts
export enum MedicalOptionalErrorType {
  required = 'required',
  invalid = 'invalid',
  duplicated = 'duplicated',
  startGreaterThanEnd = 'startGreaterThanEnd',
  smallDateInterval = 'smallDateInterval',
  createFailed = 'createFailed',
  updateFailed = 'updateFailed',
  confirmFailed = 'confirmFailed',
  isPast = 'isPast',
  incomplete = 'incomplete',
}